Mechanism to override AC_CHECK_FUNCS in Autoconf
For R packages using Autoconf, we'll sometimes need to be able to
force the result of AC_CHECK_FUNCS.

For example, for the `uuid` package Autoconf sets `HAVE_GETRANDOM`.
Despite compiling OK, using `getrandom()` at the time of writing
in fact fails for R packages compiled using Emscripten, due to a
missing symbol. Autoconf missed this because we compile R packages
with the `SIDE_MODULE=1` flag, and with this flag Emscripten does
not complain about the missing symbols.

So, using this mechanism we can force Autoconf to return "no" when
checking for `getrandom()`, fixing the issue with `uuid`.